feat: improve suite validity
